# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Port Logistics API",
        "description": "Port Logistics API",
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://port_logistic.com/",
            "description": ""
        }
    ],
    "paths": {
...
        "/users": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get list of users",
                "description": "Returns list of users",
                "operationId": "App\\Http\\Controllers\\API\\V1\\UserController::index",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/user": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Insert User",
                "operationId": "App\\Http\\Controllers\\API\\V1\\UserController::store",
                "requestBody": {
                    "description": "Pass user details",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "first_name",
                                    "last_name",
                                    "mobile_no",
                                    "address1",
                                    "address2",
                                    "country_id",
                                    "pin_code",
                                    "state_id",
                                    "city_id",
                                    "department_id",
                                    "role_id"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "email",
                                        "format": "email",
                                        "example": "admin@esspl.com"
                                    },
                                    "first_name": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Ess"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Admin"
                                    },
                                    "mobile_no": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "99999999"
                                    },
                                    "address1": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "1"
                                    },
                                    "address2": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "1"
                                    },
                                    "country_id": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "1"
                                    },
                                    "pin_code": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "751023"
                                    },
                                    "state_id": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "1"
                                    },
                                    "city_id": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "1"
                                    },
                                    "department_id": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "1"
                                    },
                                    "role_id": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/user/{id}": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get User Details By Id",
                "operationId": "App\\Http\\Controllers\\API\\V1\\UserController::edit",
                "requestBody": {
                    "description": "Pass User Id",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "Update User",
                "operationId": "App\\Http\\Controllers\\API\\V1\\UserController::update",
                "requestBody": {
                    "description": "Pass user details",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id",
                                    "email",
                                    "first_name",
                                    "last_name",
                                    "mobile_no",
                                    "address1",
                                    "address2",
                                    "country_id",
                                    "pin_code",
                                    "state_id",
                                    "city_id",
                                    "department_id",
                                    "role_id"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "1"
                                    },
                                    "email": {
                                        "type": "email",
                                        "format": "email",
                                        "example": "admin@esspl.com"
                                    },
                                    "first_name": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Ess"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Admin"
                                    },
                                    "mobile_no": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "99999999"
                                    },
                                    "address1": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "1"
                                    },
                                    "address2": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "1"
                                    },
                                    "country_id": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "1"
                                    },
                                    "pin_code": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "751023"
                                    },
                                    "state_id": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "1"
                                    },
                                    "city_id": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "1"
                                    },
                                    "department_id": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "1"
                                    },
                                    "role_id": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and mobile to get the authentication token",
                "name": "Token based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}